pyproject.toml
[project]
name = "practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.115.13",
    "openai-agents[litellm]>=0.0.19",
]
